make reusable virtualized feed component
